Add support for the new El Capitan USB stack
Apple revamped the USB stack in OS X 10.11, and dropped the 'PortNum'
attribute from the I/O registry.

The new code walks up the tree in the service plane, using the 'port'
attribute on AppleUSBHostPort entries to get the same information.

I fixed some leaks and improved the code along the way, and the commits
actually drops more lines than it adds despite the need to support two
USB stacks.